Revert "Add ItemsAdder verification for Mascot (#389)"
This reverts commit c3f5354374cc95ecc9ead21842a441e0b048e688.
