+++ src/main/java/fr/openmc/core/features/city/mascots/MascotsListener.java
@@ -1,5 +1,7 @@
 package fr.openmc.core.features.city.mascots;
 
+import dev.lone.itemsadder.api.CustomBlock;
+import dev.lone.itemsadder.api.Events.CustomBlockPlaceEvent;
 import fr.openmc.core.OMCPlugin;
 import fr.openmc.core.features.city.City;
 import fr.openmc.core.features.city.CityManager;
@@ -11,6 +13,7 @@ import fr.openmc.core.utils.chronometer.ChronometerType;
 import fr.openmc.core.utils.messages.MessageType;
 import fr.openmc.core.utils.messages.MessagesManager;
 import fr.openmc.core.utils.messages.Prefix;
+import io.papermc.paper.event.entity.EntityDamageItemEvent;
 import io.papermc.paper.event.entity.EntityMoveEvent;
 import lombok.SneakyThrows;
 import net.kyori.adventure.text.Component;
@@ -289,6 +292,21 @@ public class MascotsListener implements Listener {
 
             }
             e.setCancelled(true);
+        }
+    }
+
+    @EventHandler
+    public void onBlockPlace(CustomBlockPlaceEvent event) {
+        Block block = event.getBlock();
+        Location loc = block.getLocation();
+
+        Collection<Entity> nearbyEntities = loc.getWorld().getNearbyEntities(loc, 1.5, 1.5, 1.5);
+
+        for (Entity entity : nearbyEntities) {
+            if (MascotUtils.isMascot(entity)) {
+                event.setCancelled(true);
+                return;
+            }
         }
     }
